Yay, tamed that computer into linearity, at least somewhat..

+++ rust/temp_controller/src/iir.rs
@@ -11,9 +11,9 @@ pub type Vec5 = [i32; 5];
 
 /// Main IIR struct holds coefficient vector and a shift value which defines the fixed point position
 pub struct Iir {
-    pub ba: Vec5,   // b and a coeffitients can be changed. [a0,a1,b0,b1,b2]
+    pub ba: Vec5,   // b and a coeffitients can be changed. [b0,b1,b2,a1,a2]
     pub shift: i32, // shift for fixed point pos
-    pub xy: Vec5,       // x and y internal filter states       [x0,x1,y0,y1,y2]
+    pub xy: Vec5,   // x and y internal filter states       [x0,x1,y0,y1,y2]
 }
 
 impl Iir {
@@ -24,7 +24,7 @@ impl Iir {
         self.xy.copy_within(0..4, 1);
         self.xy[0] = x0;
 
-        let y0 = 1 << ((self.shift as i64) - 1);
+        let y0 = 1 << ((self.shift) - 1);
         let y = &self.xy
             .iter()
             .zip(&self.ba)
